Make it modular
