+++ src/index.html
@@ -9,7 +9,7 @@
     <section>
       <div id="head-list">
         <h2>Today's To Do</h2>
-        <span>&#x021BA;</span>
+        <span class="refresh">&#x021BA;</span>
       </div>
       <div id="add-list">
         <label for="input-list"></label>
